pyproject.toml
[project]
name = "ocr-document-categorizer"
version = "0.1.0"
description = "Crop documents, run OCR, and categorize text into headings."
requires-python = ">=3.9"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Your Name" }]

dependencies = [
  "opencv-python>=4.8.0",
  "numpy>=1.24.0",
  "pillow>=10.0.0",
  "tqdm>=4.66.0",
  "reportlab>=4.0.0"
]

[project.optional-dependencies]
ocr = ["easyocr>=1.7.0"]
desktop = ["tkinterdnd2>=0.4.2"]

[project.scripts]
ocr-doc-cat = "ocr_document_categorizer.cli:main"
ocr-doc-cat-desktop = "ocr_document_categorizer.app:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
